Moving PathString to its own file for extension
